+++ R/helper_functions.R
@@ -269,12 +269,12 @@
 # to reduce redundant code, I extract this part which is used several
 # times accross this package
 .get_elements <- function(effects, component) {
-  elements <- c("conditional", "nonlinear", "random", "zero_inflated", "zero_inflated_random", "dispersion", "instruments", "interactions", "simplex", "smooth_terms", "sigma", "nu", "tau", "correlation", "slopes", "cluster")
+  elements <- c("conditional", "nonlinear", "random", "zero_inflated", "zero_inflated_random", "dispersion", "instruments", "interactions", "simplex", "smooth_terms", "sigma", "nu", "tau", "correlation", "slopes", "cluster", "extra")
 
   elements <- switch(
     effects,
     all = elements,
-    fixed = elements[elements %in% c("conditional", "zero_inflated", "dispersion", "instruments", "interactions", "simplex", "smooth_terms", "correlation", "slopes", "sigma", "nonlinear", "cluster")],
+    fixed = elements[elements %in% c("conditional", "zero_inflated", "dispersion", "instruments", "interactions", "simplex", "smooth_terms", "correlation", "slopes", "sigma", "nonlinear", "cluster", "extra")],
     random = elements[elements %in% c("random", "zero_inflated_random")]
   )
 
@@ -293,7 +293,8 @@
     correlation = elements[elements == "correlation"],
     cluster = elements[elements == "cluster"],
     nonlinear = elements[elements == "nonlinear"],
-    slopes = elements[elements == "slopes"]
+    slopes = elements[elements == "slopes"],
+    extra = elements[elements == "extra"]
   )
 
   elements